Refactored association sorting  to common util
